Сделал дизайн

+++ src/components/Reg.vue
@@ -1,10 +1,58 @@
 <template>
-  <div class="register">
-    <input class="form_input" type="text" placeholder="ФИО" v-model="user.fio">
-    <input class="form_input" type="text" placeholder="Логин" v-model="user.login">
-    <input class="form_input" type="password" placeholder="Пароль" v-model="user.password">
-    <input class="form_input" type="submit" value="Зарегистрироваться" @keydown.enter="registerUser" @click="registerUser">
-  </div>
+  <v-form
+      ref="form"
+      lazy-validation
+      class="register"
+  >
+
+    <v-app-bar-title style="text-align: center; cursor: default">Регистрация</v-app-bar-title>
+
+    <v-text-field
+        v-model="user.fio"
+        label="ФИО"
+        required
+    ></v-text-field>
+
+    <v-text-field
+        v-model="user.login"
+        label="Логин"
+        required
+    ></v-text-field>
+
+    <v-text-field
+        type="password"
+        v-model="user.password"
+        label="Пароль"
+        required
+    ></v-text-field>
+
+    <v-text-field
+        v-if="user.password !== null"
+        type="password"
+        v-model="password_confirm"
+        label="Подтвердите пароль"
+        required
+    ></v-text-field>
+
+    <v-alert
+        v-if="message"
+        type="error"
+        class="mes"
+    >
+      {{message}}
+    </v-alert>
+
+    <v-btn
+        color="success"
+        @keydown.enter="registerUser"
+        @click="registerUser"
+        plain
+        type="submit"
+    >
+      Зарегистрироваться
+    </v-btn>
+
+  </v-form>
 </template>
 
 <script>
@@ -16,20 +64,36 @@ export default {
         fio: null,
         login: null,
         password: null,
-      }
+      },
+      password_confirm : '',
+      message : '',
     }
   },
   methods: {
     registerUser() {
-      let users = this.$store.state.userList;
-      users.push(this.user);
-      this.$store.commit('setUserList', users);
-      localStorage.setItem('userList', JSON.stringify(users));
-      this.$store.commit('setAuth');
-      this.$store.commit('setUser', this.user);
-      localStorage.setItem('auth', "true");
-      localStorage.setItem('user', JSON.stringify(this.user));
-      this.$router.push('/todolist');
+      if (this.user.password === null || this.user.login === null || this.user.fio === null || this.password_confirm === null) {
+        this.message = 'Все поля обязательны для заполнения';
+      }
+      else if (this.user.password === this.password_confirm) {
+        let user_exist = this.$store.state.userList.find(item => item.login === this.user.login);
+        if (user_exist != undefined) {
+          this.message = 'Такой логин занят';
+        }
+        else {
+          let users = this.$store.state.userList;
+          users.push(this.user);
+          this.$store.commit('setUserList', users);
+          localStorage.setItem('userList', JSON.stringify(users));
+          this.$store.commit('setAuth');
+          this.$store.commit('setUser', this.user);
+          localStorage.setItem('auth', "true");
+          localStorage.setItem('user', JSON.stringify(this.user));
+          this.$router.push('/todolist');
+        }
+      }
+      else {
+        this.message = 'Пароли не совпадают';
+      }
     }
   }
 }
@@ -38,7 +102,7 @@ export default {
 <style scoped>
   .register {
     margin: 0 auto;
-    width: 400px;
+    width: 720px;
     display: flex;
     flex-direction: column;
   }
